+++ testground/benchmark/benchmark/transaction.py
@@ -10,6 +10,7 @@ import aiohttp
 import backoff
 import eth_abi
 import ujson
+from eth_account._utils.legacy_transactions import Transaction
 from hexbytes import HexBytes
 
 from . import cosmostx
@@ -20,6 +21,21 @@ GAS_PRICE = 1000000000
 CHAIN_ID = 777
 CONNECTION_POOL_SIZE = 1024
 TXS_DIR = "txs"
+
+Job = namedtuple(
+    "Job",
+    [
+        "chunk",
+        "global_seq",
+        "num_txs",
+        "tx_type",
+        "create_tx",
+        "batch",
+        "nonce",
+        "msg_version",
+    ],
+)
+EthTx = namedtuple("EthTx", ["tx", "raw", "sender"])
 
 
 def simple_transfer_tx(sender: str, nonce: int):
@@ -53,11 +69,48 @@ TX_TYPES = {
 }
 
 
-Job = namedtuple(
-    "Job",
-    ["chunk", "global_seq", "num_accounts", "num_txs", "tx_type", "create_tx", "batch"],
-)
-EthTx = namedtuple("EthTx", ["tx", "raw", "sender"])
+def build_evm_msg_1_3(tx: EthTx):
+    """
+    build cronos v1.3 version of MsgEthereumTx
+    """
+    txn = Transaction.from_bytes(tx.raw)
+    return cosmostx.build_any(
+        cosmostx.MsgEthereumTx.MSG_URL,
+        cosmostx.MsgEthereumTx(
+            data=cosmostx.build_any(
+                cosmostx.LegacyTx.MSG_URL,
+                cosmostx.LegacyTx(
+                    nonce=txn.nonce,
+                    gas_price=str(txn.gasPrice),
+                    gas=txn.gas,
+                    to=txn.to.hex(),
+                    value=str(txn.value),
+                    data=txn.data,
+                    v=txn.v.to_bytes(32, byteorder="big"),
+                    r=txn.r.to_bytes(32, byteorder="big"),
+                    s=txn.s.to_bytes(32, byteorder="big"),
+                ),
+            ),
+            deprecated_hash=txn.hash().hex(),
+            from_=tx.sender,
+        ),
+    )
+
+
+def build_evm_msg_1_4(tx: EthTx):
+    return cosmostx.build_any(
+        cosmostx.MsgEthereumTx.MSG_URL,
+        cosmostx.MsgEthereumTx(
+            from_=tx.sender,
+            raw=tx.raw,
+        ),
+    )
+
+
+MSG_VERSIONS = {
+    "1.3": build_evm_msg_1_3,
+    "1.4": build_evm_msg_1_4,
+}
 
 
 def _do_job(job: Job):
@@ -67,7 +120,7 @@ def _do_job(job: Job):
     for acct in accounts:
         txs = []
         for i in range(job.num_txs):
-            tx = job.create_tx(acct.address, i)
+            tx = job.create_tx(acct.address, job.nonce + i)
             raw = acct.sign_transaction(tx).rawTransaction
             txs.append(EthTx(tx, raw, HexBytes(acct.address)))
             total += 1
@@ -76,19 +129,37 @@ def _do_job(job: Job):
 
         # to keep it simple, only build batch inside the account
         txs = [
-            build_cosmos_tx(*txs[start:end])
+            build_cosmos_tx(*txs[start:end], msg_version=job.msg_version)
             for start, end in split_batch(len(txs), job.batch)
         ]
         acct_txs.append(txs)
     return acct_txs
 
 
-def gen(global_seq, num_accounts, num_txs, tx_type: str, batch: int) -> [str]:
+def gen(
+    global_seq,
+    num_accounts,
+    num_txs,
+    tx_type: str,
+    batch: int,
+    nonce: int = 0,
+    start_account: int = 0,
+    msg_version: str = "1.4",
+) -> [str]:
     chunks = split(num_accounts, os.cpu_count())
     create_tx = TX_TYPES[tx_type]
     jobs = [
-        Job(chunk, global_seq, num_accounts, num_txs, tx_type, create_tx, batch)
-        for chunk in chunks
+        Job(
+            (start + start_account, end + start_account),
+            global_seq,
+            num_txs,
+            tx_type,
+            create_tx,
+            batch,
+            nonce,
+            msg_version,
+        )
+        for start, end in chunks
     ]
 
     with multiprocessing.Pool() as pool:
@@ -119,20 +190,12 @@ def load(datadir: Path, global_seq: int) -> [str]:
         return ujson.load(f)
 
 
-def build_cosmos_tx(*txs: EthTx) -> str:
+def build_cosmos_tx(*txs: EthTx, msg_version="1.4", evm_denom=DEFAULT_DENOM) -> str:
     """
     return base64 encoded cosmos tx, support batch
     """
-    msgs = [
-        cosmostx.build_any(
-            "/ethermint.evm.v1.MsgEthereumTx",
-            cosmostx.MsgEthereumTx(
-                from_=tx.sender,
-                raw=tx.raw,
-            ),
-        )
-        for tx in txs
-    ]
+    build_msg = MSG_VERSIONS[msg_version]
+    msgs = [build_msg(tx) for tx in txs]
     fee = sum(tx.tx["gas"] * tx.tx["gasPrice"] for tx in txs)
     gas = sum(tx.tx["gas"] for tx in txs)
     body = cosmostx.TxBody(
@@ -143,7 +206,7 @@ def build_cosmos_tx(*txs: EthTx) -> str:
     )
     auth_info = cosmostx.AuthInfo(
         fee=cosmostx.Fee(
-            amount=[cosmostx.Coin(denom=DEFAULT_DENOM, amount=str(fee))],
+            amount=[cosmostx.Coin(denom=evm_denom, amount=str(fee))],
             gas_limit=gas,
         )
     )
@@ -154,20 +217,19 @@ def build_cosmos_tx(*txs: EthTx) -> str:
     ).decode()
 
 
+def json_rpc_send_body(raw, method="broadcast_tx_async"):
+    return {
+        "jsonrpc": "2.0",
+        "method": method,
+        "params": {"tx": raw},
+        "id": 1,
+    }
+
+
 @backoff.on_predicate(backoff.expo, max_time=60, max_value=5)
 @backoff.on_exception(backoff.expo, aiohttp.ClientError, max_time=60, max_value=5)
-async def async_sendtx(session, raw):
-    async with session.post(
-        LOCAL_RPC,
-        json={
-            "jsonrpc": "2.0",
-            "method": "broadcast_tx_async",
-            "params": {
-                "tx": raw,
-            },
-            "id": 1,
-        },
-    ) as rsp:
+async def async_sendtx(session, raw, rpc):
+    async with session.post(rpc, json=json_rpc_send_body(raw)) as rsp:
         data = await rsp.json()
         if "error" in data:
             print("send tx error, will retry,", data["error"])
@@ -175,10 +237,10 @@ async def async_sendtx(session, raw):
         return True
 
 
-async def send(txs):
+async def send(txs, rpc=LOCAL_RPC):
     connector = aiohttp.TCPConnector(limit=CONNECTION_POOL_SIZE)
     async with aiohttp.ClientSession(
         connector=connector, json_serialize=ujson.dumps
     ) as session:
-        tasks = [asyncio.ensure_future(async_sendtx(session, raw)) for raw in txs]
+        tasks = [asyncio.ensure_future(async_sendtx(session, raw, rpc)) for raw in txs]
         await asyncio.gather(*tasks)